Fix to bug where grade weighting was cast as an int on output, causing it to be saved as the int value if not fixed before editing the assessment. Additionally, csv import now working with variety of inputs in the CSV file based on the marking scheme for the assessment being imported to, rather than just using percentages all the time.

+++ www-root/api/assessment-weighting-exception.api.php
@@ -58,7 +58,7 @@ if((!isset($_SESSION["isAuthorized"])) || (!$_SESSION["isAuthorized"])) {
 	}
 	
 	if (isset($_POST["grade_weighting"])) {
-		$grade_weighting = clean_input($_POST["grade_weighting"], "int");
+		$grade_weighting = clean_input($_POST["grade_weighting"], "float");
 	}
 	
 	if (isset($_POST["remove"]) && $_POST["remove"] == "1") {
